test: organize learner chat hook helpers
Changed:
Moved shared useChatLogicHook test wrappers, deferred helpers, base params, and mock run-source helpers into a dedicated test utility module.

Benefit:
Learner chat hook tests are easier to scan and maintain without changing covered behavior.
