+++ src/cook-web/src/app/c/[[...id]]/Components/ChatUi/useChatLogicHook.test.tsx
@@ -1,9 +1,16 @@
-import React from 'react';
 import { act, renderHook, waitFor } from '@testing-library/react';
 import { toast, toastOnce } from '@/hooks/useToast';
 import useChatLogicHook, { ChatContentItemType } from './useChatLogicHook';
-import { AppContext } from '../AppContext';
 import { SSE_INPUT_TYPE, SSE_OUTPUT_TYPE } from '@/c-api/studyV2';
+import {
+  buildBaseParams,
+  createDeferred,
+  mobileWrapper,
+  MockRunSource,
+  wrapper,
+  type ActiveRun,
+  type RunRequestBody,
+} from './useChatLogicHook.testUtils';
 import { stopAllActiveLessonStreams } from '@/app/c/[[...id]]/events';
 import { useLessonRunContentStore } from '@/c-store/useLessonRunContentStore';
 
@@ -171,39 +178,8 @@ jest.mock('@/c-api/studyV2', () => {
   };
 });
 
-type Listener = (event?: Event) => void;
-
-class MockRunSource {
-  readyState = 0;
-
-  private listeners = new Map<string, Listener[]>();
-
-  addEventListener = jest.fn((type: string, listener: Listener) => {
-    const existing = this.listeners.get(type) ?? [];
-    existing.push(listener);
-    this.listeners.set(type, existing);
-  });
-
-  close = jest.fn(() => {
-    this.readyState = 2;
-    this.emit('readystatechange');
-  });
-
-  emit(type: string, event?: Event) {
-    for (const listener of this.listeners.get(type) ?? []) {
-      listener(event);
-    }
-  }
-}
-
 describe('useChatLogicHook stream cleanup', () => {
-  let activeRun:
-    | {
-        source: MockRunSource;
-        onMessage: (response: any) => Promise<void> | void;
-        onError: (error: unknown) => void;
-      }
-    | undefined;
+  let activeRun: ActiveRun | undefined;
 
   beforeEach(() => {
     jest.clearAllMocks();
@@ -232,10 +208,7 @@ describe('useChatLogicHook stream cleanup', () => {
         _shifuBid: string,
         _outlineBid: string,
         _previewMode: boolean,
-        _body: {
-          input: string | Record<string, any>;
-          input_type: SSE_INPUT_TYPE;
-        },
+        _body: RunRequestBody,
         onMessage: (response: any) => Promise<void> | void,
         onError: (error: unknown) => void,
       ) => {
@@ -248,59 +221,6 @@ describe('useChatLogicHook stream cleanup', () => {
         return source;
       },
     );
-  });
-
-  const wrapper = ({ children }: { children: React.ReactNode }) => (
-    <AppContext.Provider
-      value={{
-        isLoggedIn: false,
-        mobileStyle: false,
-        userInfo: null,
-        theme: 'light',
-        frameLayout: 0,
-      }}
-    >
-      {children}
-    </AppContext.Provider>
-  );
-
-  const mobileWrapper = ({ children }: { children: React.ReactNode }) => (
-    <AppContext.Provider
-      value={{
-        isLoggedIn: false,
-        mobileStyle: true,
-        userInfo: null,
-        theme: 'light',
-        frameLayout: 0,
-      }}
-    >
-      {children}
-    </AppContext.Provider>
-  );
-
-  const createDeferred = <T,>() => {
-    let resolve!: (value: T) => void;
-    const promise = new Promise<T>(next => {
-      resolve = next;
-    });
-    return { promise, resolve };
-  };
-
-  const buildBaseParams = () => ({
-    shifuBid: 'shifu-1',
-    outlineBid: 'lesson-1',
-    lessonId: 'lesson-1',
-    lessonHasContentUpdate: false,
-    trackEvent: jest.fn(),
-    trackTrailProgress: jest.fn(),
-    lessonUpdate: jest.fn(),
-    chapterUpdate: jest.fn(),
-    updateSelectedLesson: jest.fn(),
-    getNextLessonId: jest.fn(() => null),
-    scrollToLesson: jest.fn(),
-    showOutputInProgressToast: jest.fn(),
-    onPayModalOpen: jest.fn(),
-    onGoChapter: jest.fn(),
   });
 
   it('sends listen=false in the run body when listen requests are disabled', async () => {
